+++ src/main/java/com/debateseason_backend_v1/domain/chat/presentation/controller/ChatControllerV1.java
@@ -1,9 +1,9 @@
-package com.debateseason_backend_v1.domain.chat.controller;
+package com.debateseason_backend_v1.domain.chat.presentation.controller;
 
 import com.debateseason_backend_v1.common.response.ApiResult;
 import com.debateseason_backend_v1.common.response.ErrorResponse;
-import com.debateseason_backend_v1.domain.chat.model.request.ChatReactionRequest;
-import com.debateseason_backend_v1.domain.chat.model.response.ChatMessagesResponse;
+import com.debateseason_backend_v1.domain.chat.presentation.dto.request.ChatReactionRequest;
+import com.debateseason_backend_v1.domain.chat.presentation.dto.response.ChatMessagesResponse;
 import com.debateseason_backend_v1.domain.chat.application.service.ChatReactionServiceV1;
 import com.debateseason_backend_v1.security.jwt.JwtUtil;
 import lombok.extern.slf4j.Slf4j;
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 
 import com.debateseason_backend_v1.domain.chat.application.service.ChatServiceV1;
-import com.debateseason_backend_v1.domain.chat.model.response.ChatMessageResponse;
+import com.debateseason_backend_v1.domain.chat.presentation.dto.response.ChatMessageResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
