+++ src/main/java/com/debateseason_backend_v1/domain/chat/presentation/controller/WebSocketControllerV1.java
@@ -1,12 +1,12 @@
-package com.debateseason_backend_v1.domain.chat.controller;
+package com.debateseason_backend_v1.domain.chat.presentation.controller;
 
 import java.time.LocalDateTime;
 
 import com.debateseason_backend_v1.common.exception.CustomException;
-import com.debateseason_backend_v1.domain.chat.model.response.ChatMessageErrorResponse;
-import com.debateseason_backend_v1.domain.chat.model.response.ChatMessageResponse;
+import com.debateseason_backend_v1.domain.chat.presentation.dto.response.ChatMessageErrorResponse;
+import com.debateseason_backend_v1.domain.chat.presentation.dto.response.ChatMessageResponse;
 import com.debateseason_backend_v1.domain.chat.application.service.ChatServiceV1;
-import com.debateseason_backend_v1.domain.chat.valide.ChatValidate;
+import com.debateseason_backend_v1.domain.chat.validation.ChatValidate;
 import jakarta.validation.Valid;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.messaging.handler.annotation.*;
@@ -17,7 +17,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 
 import com.debateseason_backend_v1.common.enums.MessageType;
-import com.debateseason_backend_v1.domain.chat.model.request.ChatMessageRequest;
+import com.debateseason_backend_v1.domain.chat.presentation.dto.request.ChatMessageRequest;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
