+++ src/main/java/com/debateseason_backend_v1/domain/chat/presentation/dto/request/ChatReactionRequest.java
@@ -1,4 +1,4 @@
-package com.debateseason_backend_v1.domain.chat.model.request;
+package com.debateseason_backend_v1.domain.chat.presentation.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
